Import vagrant boxes.
